fixed win rate calucation
