+++ chart_dash.py
@@ -62,7 +62,7 @@ app = Dash(__name__, suppress_callback_exceptions=True)
 # Layout mit Darkmode CSS Styles
 app.layout = html.Div(
     style={
-        "backgroundColor": "#1e1e1e",
+        "backgroundColor": "#2f2f2f",
         "color": "white",
         "font-family": "Arial, sans-serif",
         "padding": "20px",
@@ -136,11 +136,23 @@ app.layout = html.Div(
                     multi=True,
                     value=[],
                     placeholder="Select period",
-                    style={"width": "50%", "color": "#000"},
+                    style={"width": "400px", "color": "#000"},
                 )
             ],
         ),
         dcc.Graph(id="combined-chart"),
+        html.Div(
+            style={
+                "display": "flex",
+                "flexDirection": "row",
+                "gap": "20px",
+                "marginTop": "20px",
+            },
+            children=[
+                dcc.Graph(id="win-rate-indicator", style={"width": "50%"}),
+                dcc.Graph(id="pcr-indicator", style={"width": "50%"}),
+            ],
+        ),
     ],
 )
 
@@ -179,17 +191,21 @@ def manage_time_period_dropdown(selected_time_filter):
     return style, options, value, placeholder
 
 
-# Callback zur Aktualisierung des kombinierten Graphen
+# Callback zur Aktualisierung des kombinierten Graphen und der Indikatoren
 @app.callback(
-    Output("combined-chart", "figure"),
+    [
+        Output("combined-chart", "figure"),
+        Output("win-rate-indicator", "figure"),
+        Output("pcr-indicator", "figure"),
+    ],
     [
         Input("time-filter", "value"),
         Input("asset-filter", "value"),
         Input("symbol-filter", "value"),
         Input("time-period-dropdown", "value"),
-    ],  # Dieser Input ist jetzt immer gültig
+    ],
 )
-def update_combined_chart(
+def update_combined_chart_and_indicators(
     time_filter, selected_assets, selected_symbols, selected_periods
 ):
     filtered_df = df.copy()
@@ -204,11 +220,8 @@ def update_combined_chart(
     if time_filter == "Total":
         pass
     elif time_filter in ["All weeks", "All months", "All quarters"]:
-        # selected_periods ist [] wenn nichts ausgewählt oder Dropdown gerade erst sichtbar wurde/Filtertyp änderte
         if not selected_periods:
             filtered_df = pd.DataFrame(columns=df.columns)
-            if "tradeDate" in filtered_df.columns:
-                filtered_df["tradeDate"] = pd.to_datetime(filtered_df["tradeDate"])
         else:
             if time_filter == "All weeks":
                 filtered_df["week_str"] = filtered_df["tradeDate"].dt.strftime("%Y-W%U")
@@ -230,13 +243,76 @@ def update_combined_chart(
                     filtered_df["quarter_str"].isin(selected_periods)
                 ]
 
+    # Win-Rate Berechnung
+    if not filtered_df.empty:
+        trade_pnl_sum = filtered_df.groupby("opendateTime")["PnLRealized"].sum()
+        wins = (trade_pnl_sum > 0).sum()
+        total_trades = len(trade_pnl_sum)
+        win_rate = (wins / total_trades * 100) if total_trades > 0 else 0
+    else:
+        win_rate = 0
+        wins = 0
+        total_trades = 0
+
+    win_rate_figure = go.Figure(
+        go.Indicator(
+            mode="gauge+number",
+            value=win_rate,
+            title={"text": "Win Rate (%)"},
+            gauge={
+                "axis": {
+                    "range": [0, 100],
+                    "tickcolor": "white",
+                    "tickwidth": 1,
+                },
+                "bordercolor": "white",
+                "borderwidth": 1,
+            },
+        )
+    )
+    win_rate_figure.update_layout(
+        plot_bgcolor="#3f3f3f",
+        paper_bgcolor="#3f3f3f",
+        font=dict(color="white"),
+    )
+
+    # PCR Berechnung
+    pcr_df = filtered_df[filtered_df["assetCategory"].isin(["FOP", "OPT"])]
+    if not pcr_df.empty:
+        sum_cost = abs(pcr_df["cost"].sum()).round(2)
+        sum_realized = pcr_df["PnLRealized"].sum().round(2)
+        pcr = (sum_realized / sum_cost * 100).round(2) if sum_cost != 0 else 0
+    else:
+        pcr = 0
+
+    pcr_figure = go.Figure(
+        go.Indicator(
+            mode="gauge+number",
+            value=pcr,
+            title={"text": "PCR (%)"},
+            gauge={
+                "axis": {
+                    "range": [0, 100],
+                    "tickcolor": "white",
+                    "tickwidth": 1,
+                },
+                "bordercolor": "white",
+                "borderwidth": 1,
+            },
+        )
+    )
+    pcr_figure.update_layout(
+        plot_bgcolor="#3f3f3f", paper_bgcolor="#3f3f3f", font=dict(color="white")
+    )
+
+    # Chart-Berechnung (bestehender Code)
     if filtered_df.empty:
-        return {
+        combined_chart = {
             "data": [],
             "layout": go.Layout(
                 title="No data for the selected filters",
-                plot_bgcolor="#1e1e1e",
-                paper_bgcolor="#1e1e1e",
+                plot_bgcolor="#3f3f3f",
+                paper_bgcolor="#3f3f3f",
                 font=dict(color="white"),
                 xaxis=dict(
                     showgrid=False,
@@ -247,66 +323,69 @@ def update_combined_chart(
                 yaxis=dict(showgrid=False, zeroline=False, showticklabels=False),
             ),
         }
-
-    pnl_per_day = (
-        filtered_df.groupby(filtered_df["tradeDate"].dt.date)["PnLRealized"]
-        .sum()
-        .reset_index()
-    )
-    pnl_per_day["tradeDate"] = pd.to_datetime(pnl_per_day["tradeDate"])
-    pnl_per_day = pnl_per_day.sort_values("tradeDate")
-    pnl_per_day["TotalProfit"] = pnl_per_day["PnLRealized"].cumsum()
-    pnl_per_day["tradeDateStr"] = pnl_per_day["tradeDate"].dt.strftime("%Y-%m-%d")
-
-    fig = go.Figure()
-    fig.add_trace(
-        go.Bar(
-            x=pnl_per_day["tradeDateStr"],
-            y=pnl_per_day["PnLRealized"],
-            name="Daily PnL",
-            marker_color=[
-                "crimson" if x < 0 else "lightseagreen"
-                for x in pnl_per_day["PnLRealized"]
-            ],
+    else:
+        pnl_per_day = (
+            filtered_df.groupby(filtered_df["tradeDate"].dt.date)["PnLRealized"]
+            .sum()
+            .reset_index()
         )
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=pnl_per_day["tradeDateStr"],
-            y=pnl_per_day["TotalProfit"],
-            mode="lines+markers",
-            name="Cumulative Total Profit",
-            line=dict(color="orange", width=2),
-            marker=dict(size=5),
+        pnl_per_day["tradeDate"] = pd.to_datetime(pnl_per_day["tradeDate"])
+        pnl_per_day = pnl_per_day.sort_values("tradeDate")
+        pnl_per_day["TotalProfit"] = pnl_per_day["PnLRealized"].cumsum()
+        pnl_per_day["tradeDateStr"] = pnl_per_day["tradeDate"].dt.strftime("%Y-%m-%d")
+
+        combined_chart = go.Figure()
+        combined_chart.add_trace(
+            go.Bar(
+                x=pnl_per_day["tradeDateStr"],
+                y=pnl_per_day["PnLRealized"],
+                name="Daily PnL",
+                marker_color=[
+                    "crimson" if x < 0 else "lightseagreen"
+                    for x in pnl_per_day["PnLRealized"]
+                ],
+            )
         )
-    )
-    fig.update_layout(
-        title="Daily PnL and Cumulative Total Profit",
-        xaxis=dict(
-            title="Date",
-            type="category",
-            showgrid=True,
-            gridcolor="rgba(128,128,128,0.2)",
-            color="white",
-            tickangle=-45,
-        ),
-        yaxis=dict(
-            title="Amount",
-            showgrid=True,
-            gridcolor="rgba(128,128,128,0.2)",
-            color="white",
-            zerolinecolor="gray",
-            zerolinewidth=1,
-        ),
-        plot_bgcolor="#1e1e1e",
-        paper_bgcolor="#1e1e1e",
-        font=dict(color="white"),
-        legend=dict(
-            bgcolor="rgba(30,30,30,0.7)", bordercolor="gray", font=dict(color="white")
-        ),
-        margin=dict(l=60, r=40, t=80, b=100),
-    )
-    return fig
+        combined_chart.add_trace(
+            go.Scatter(
+                x=pnl_per_day["tradeDateStr"],
+                y=pnl_per_day["TotalProfit"],
+                mode="lines+markers",
+                name="Cumulative Total Profit",
+                line=dict(color="orange", width=2),
+                marker=dict(size=5),
+            )
+        )
+        combined_chart.update_layout(
+            title="Daily PnL and Cumulative Total Profit",
+            xaxis=dict(
+                title="Date",
+                type="category",
+                showgrid=True,
+                gridcolor="rgba(128,128,128,0.2)",
+                color="white",
+                tickangle=-45,
+            ),
+            yaxis=dict(
+                title="Amount",
+                showgrid=True,
+                gridcolor="rgba(128,128,128,0.2)",
+                color="white",
+                zerolinecolor="gray",
+                zerolinewidth=1,
+            ),
+            plot_bgcolor="#3f3f3f",
+            paper_bgcolor="#3f3f3f",
+            font=dict(color="white"),
+            legend=dict(
+                bgcolor="#3f3f3f",
+                bordercolor="gray",
+                font=dict(color="white"),
+            ),
+            margin=dict(l=60, r=40, t=80, b=100),
+        )
+
+    return combined_chart, win_rate_figure, pcr_figure
 
 
 if __name__ == "__main__":
